fix: not update length possession_list in plist statusline
The problem is when we delete a session from possession list (plist)
window, the total session in the statusline possession list window is
not updated.

This change make sure that the length of possession list is re-evaluated
if there's any changes to the value.

Also, make sure to abort the function set_options function if there's an
error.